+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [v0.2.5] - 2026-06-16
+
+New round of components
+
+### Added
+
+- Alerts, Modals, and a Promo Bar added to the core components
+
 ## [v0.2.4] - 2026-06-16
 
 Navigation enhancements to starter kit